Move JavaScript unit registry indexes
